requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
peft>=0.4.0
bitsandbytes>=0.39.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
scikit-learn>=1.3.0
rouge-score>=0.1.2
nltk>=3.8.0
jupyter>=1.0.0
ipywidgets>=8.0.0
tensorboard>=2.13.0

